Create knightMoves function
Decided to use queue, even though the implementation is inefficient,
because using a stack is going to be slower.
An array of all moves made by the moment is added to the queue. This is
the only way I managed to come up with which is able to print the path.